Add Firebase Realtime Database
Add Firebase. No longer need to store notes in the app, only the note ID needs to be temporary stored to determine the selected ID. Use "onValue" to dynamically display the notes. Use "push" to push the note to the DB. Use "get" to get the title, text, and color of a note for the given ID in the DB. Use "update" to update the note in the DB. Use "remove" to remove the note from the DB. No need to render or save the notes, since the notes are dynamically displayed whenever the DB changes.